pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ade-toolkit"
version = "0.1.0"
description = "Agentic Development Environment — portable AI-driven SDLC toolkit"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.15",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "crewai>=0.100",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-tmp-files>=0.0.2",
    "ruff>=0.4",
]

[project.scripts]
ade = "ade.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ade"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
